# search-manage/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/index/getList": {
            "get": {
                "description": "Get a list of all loaded indices",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Index"
                ],
                "summary": "Get a list of all loadded indices",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.VsIndexList"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/main.VsDefaultResponse"
                        }
                    }
                }
            }
        },
        "/index/getListUnloaded": {
            "get": {
                "description": "Get a list of all loaded indices",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Index"
                ],
                "summary": "Get a list of all loadded indices",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.VsUnloadedIndexList"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/main.VsDefaultResponse"
                        }
                    }
                }
            }
        },
        "/index/load": {
            "post": {
                "description": "Load index from DB to memory",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Index"
                ],
                "summary": "Load index",
                "parameters": [
                    {
                        "description": "Default request",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.VsDefaultRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Loaded index",
                        "schema": {
                            "$ref": "#/definitions/main.VsDefaultResponse"
                        }
                    },
                    "400": {
                        "description": "Name parameter required",
                        "schema": {
                            "$ref": "#/definitions/main.VsDefaultResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/main.VsDefaultResponse"
                        }
                    }
                }
            }
        },
        "/index/search": {
            "post": {
                "description": "Find neighbor items for a given feature vector",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Index"
                ],
                "summary": "Find neighbor items for a given feature vector",
                "parameters": [
                    {
                        "description": "Search request",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.VsSearchRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Found neighbors close to the input feature vector",
                        "schema": {
                            "$ref": "#/definitions/main.VsSearchResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.VsSearchResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/main.VsSearchResponse"
                        }
                    }
                }
            }
        },
        "/index/unload": {
            "post": {
                "description": "Unload index from memory",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Index"
                ],
                "summary": "Unload index",
                "parameters": [
                    {
                        "description": "Default request",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.VsDefaultRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Unloaded index",
                        "schema": {
                            "$ref": "#/definitions/main.VsDefaultResponse"
                        }
                    },
                    "400": {
                        "description": "Name parameter required",
                        "schema": {
                            "$ref": "#/definitions/main.VsDefaultResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/main.VsDefaultResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "main.VsDefaultRequest": {
            "type": "object",
            "properties": {
                "indexName": {
                    "type": "string"
                }
            }
        },
        "main.VsDefaultResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "main.VsIndexList": {
            "type": "object",
            "properties": {
                "dim": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "indexName": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "numVectors": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "main.VsSearchRequest": {
            "type": "object",
            "properties": {
                "dim": {
                    "type": "integer"
                },
                "indexName": {
                    "type": "string"
                },
                "numNeighbors": {
                    "type": "integer"
                },
                "numQueryVectors": {
                    "type": "integer"
                },
                "vecData": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                }
            }
        },
        "main.VsSearchResponse": {
            "type": "object",
            "properties": {
                "D": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                },
                "I": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "message": {
                    "type": "string"
                },
                "numNeighbors": {
                    "type": "integer"
                },
                "numQueryVectors": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "main.VsUnloadedIndexList": {
            "type": "object",
            "properties": {
                "indexName": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "numVectors": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        }
    }
}